[Housekeeping] Just on Curiosity in the Package Structure
