--- app/src/main/java/com/hello/curiosity/curiosity/ui/scenes/color/ColorPresentation.kt
@@ -1,8 +1,0 @@
-package com.hello.curiosity.curiosity.ui.scenes.color
-
-import androidx.compose.ui.graphics.Color
-
-data class ColorPresentation(
-    val title: String,
-    val color: Color
-)
