--- app/src/main/java/com/hello/curiosity/curiosity/ui/scenes/type/TypographyPresentation.kt
@@ -1,8 +1,0 @@
-package com.hello.curiosity.curiosity.ui.scenes.type
-
-import androidx.compose.ui.text.TextStyle
-
-data class TypographyPresentation(
-    val title: String,
-    val style: TextStyle
-)
