+++ app/src/main/java/com/hello/curiosity/ui/scenes/type/TypographyPresentation.kt
@@ -1,0 +1,8 @@
+package com.hello.curiosity.ui.scenes.type
+
+import androidx.compose.ui.text.TextStyle
+
+data class TypographyPresentation(
+    val title: String,
+    val style: TextStyle
+)
