+++ app/src/test/java/com/hello/curiosity/MainActivityTest.kt
@@ -1,0 +1,21 @@
+package com.hello.curiosity
+
+import androidx.lifecycle.Lifecycle
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+
+@RunWith(RobolectricTestRunner::class)
+class MainActivityTest {
+
+    @get:Rule
+    val activityScenarioRule = ActivityScenarioRule(MainActivity::class.java)
+
+    @Test
+    fun testActivityExists() {
+        val activity = activityScenarioRule.scenario
+        activity.moveToState(Lifecycle.State.CREATED)
+    }
+}
